refactor(domain): split academic identifiers to meet LOC limits
Move OpenAlexId, SemanticScholarId, ISSN, and ORCID Value Objects
from publications.py to a new academic_ids.py file to comply with
the 300 LOC limit for domain layer files.

- publications.py: 194 LOC (DOI, PubMedId)
- academic_ids.py: 254 LOC (OpenAlexId, SemanticScholarId, ISSN, ORCID)
